refactor(impact): promote app/impact as the neutral analysis layer, break the analyysikeskus cycle (#860)
Promote app/docs/impact/ -> top-level app/impact/ and make the
dependency graph one-directional (analyysikeskus -> impact,
docs -> impact, never the reverse).

What moved where
----------------
- git mv app/docs/impact -> app/impact (analyzer, scoring, queries,
  masking, eu_transposition; package docstring + internal :mod: refs
  repointed to app.impact.*).
- git mv app/analyysikeskus/burden.py + sanctions.py -> app/impact/
  (the full SPARQL data/query/label layer — these modules had zero
  UI/workflow code). The analyysikeskus paths are now thin re-export
  shims (from app.impact.<mod> import …) so every existing importer and
  unittest.mock patch path (app.analyysikeskus.burden.X /
  app.analyysikeskus.sanctions.X) keeps working unchanged.

Burden/sanctions split line
---------------------------
data + SPARQL query builders + Estonian label helpers (BurdenRow,
BurdenSummary, BurdenDelta, burden_label/description/key_order,
list_burden_*, burden_delta_for_draft, top_duty_holders,
bucket_burden_rows, all _build_*_query; SanctionRow, list_sanctions_*,
find_similar_sanctions, sanction_type_label/unit_label,
_xsd_decimal_literal, the _*_QUERY constants) -> app/impact.
The Halduskoormus / Sanktsioonide-indeks WORKFLOW (routes, result-shell
rendering, Estonian workflow copy, risk-band UI) stays in
app/analyysikeskus and now imports FROM app.impact.

Cycle broken
------------
The 6 function-local cycle-breaking imports in the analyzer (burden +
sanctions helpers) and the analyzer's function-local masking import are
now normal top-level imports; the TYPE_CHECKING-only BurdenRow/SanctionRow
guard is gone. The masking local import in analyysikeskus/routes.py (import
line only) is repointed too. app.impact no longer imports any consumer
layer, so app.impact.analyzer imports in a fresh interpreter with zero
app.analyysikeskus in sys.modules.

Import-direction rules now enforced (tests/test_impact_import_direction.py)
--------------------------------------------------------------------------
(a) app/docs/impact no longer exists (no shim — the path is dead);
(b) AST scan: no module under app/impact imports app.analyysikeskus /
    app.docs / app.chat / app.explorer / app.drafter / app.admin;
(c) fresh-subprocess import of app.impact.analyzer does not pull in
    app.analyysikeskus (cycle truly gone).

All app/ + tests/ docs.impact references repointed to app.impact
(grep -rn "docs\.impact" app tests == 0).

Tests
-----
ruff check app tests: clean. ruff format: clean. pyright (whole repo):
0 errors, 0 warnings. pytest full suite: 4920 passed, 50 skipped, 0 failed
(no regression vs the ~4724 baseline).

Part of #860.
